refactor(serverless-contracts): break up lampdaHandler type file
Adding the new helper type makes the file go over 200 lines.
I put the apiGateway types together, as well as some internal
authorizerContext type in their respective files, while only exporting
what was exported before.
