+++ packages/serverless-contracts/src/contracts/apiGateway/types/index.ts
@@ -3,3 +3,4 @@ export * from './requestParameters';
 export * from './fullContract';
 export * from './lambdaHandler';
 export * from './lambdaTrigger';
+export * from './apiGateway';
