+++ packages/serverless-contracts/src/contracts/apiGateway/types/lambdaHandler.ts
@@ -1,20 +1,12 @@
 import type {
-  APIGatewayEventRequestContextJWTAuthorizer,
-  APIGatewayEventRequestContextLambdaAuthorizer,
   APIGatewayEventRequestContextV2WithAuthorizer,
   APIGatewayEventRequestContextWithAuthorizer,
-  APIGatewayProxyCallback,
-  APIGatewayProxyCallbackV2,
-  APIGatewayProxyCognitoAuthorizer,
-  APIGatewayProxyEventBase,
-  APIGatewayProxyEventV2WithRequestContext,
-  APIGatewayProxyResult,
-  APIGatewayProxyResultV2,
   Callback,
   Context,
 } from 'aws-lambda';
 
 import { GenericApiGatewayContract } from '../apiGatewayContract';
+import { AuthorizerContext } from './authorizerContext';
 import {
   BodyType,
   CustomRequestContextType,
@@ -28,16 +20,6 @@ import {
   ApiGatewayIntegrationType,
 } from './constants';
 import { DefinedProperties } from './utils';
-
-type AuthorizerContext<AuthorizerType extends ApiGatewayAuthorizerType> =
-  AuthorizerType extends 'cognito'
-    ? APIGatewayProxyCognitoAuthorizer
-    : AuthorizerType extends 'jwt'
-    ? APIGatewayEventRequestContextJWTAuthorizer
-    : AuthorizerType extends 'lambda'
-    ? // We use unknown for now because we would need another schema to define the authorizer context
-      APIGatewayEventRequestContextLambdaAuthorizer<unknown>
-    : undefined;
 
 export type RequestContext<
   IntegrationType extends ApiGatewayIntegrationType,
@@ -71,10 +53,6 @@ export type HandlerEventType<
   headers: Headers;
   body: Body;
 }>;
-
-type HandlerCallback<IntegrationType> = IntegrationType extends 'restApi'
-  ? APIGatewayProxyCallback
-  : APIGatewayProxyCallbackV2;
 
 /**
  * The **internal** type of a Swarmion handler, with type-inferred event
@@ -134,51 +112,3 @@ export type SwarmionApiGatewayHandler<
   Output,
   AdditionalArgs
 >;
-
-/**
- * The type of an ApiGateway event. This is the actual event that will
- * be passed to the lambda, not the Swarmion inferred one.
- *
- * See https://docs.aws.amazon.com/lambda/latest/dg/typescript-handler.html.
- */
-export type ApiGatewayEvent<
-  IntegrationType extends ApiGatewayIntegrationType,
-  AuthorizerType extends ApiGatewayAuthorizerType,
-> = IntegrationType extends 'restApi'
-  ? APIGatewayProxyEventBase<AuthorizerContext<AuthorizerType>>
-  : APIGatewayProxyEventV2WithRequestContext<
-      APIGatewayEventRequestContextV2WithAuthorizer<
-        AuthorizerContext<AuthorizerType>
-      >
-    >;
-
-/**
- * The type of an ApiGateway event. This is the actual event that will
- * be passed to the lambda, not the Swarmion inferred one.
- *
- * See https://docs.aws.amazon.com/lambda/latest/dg/typescript-handler.html.
- */
-export type ApiGatewayResult<
-  IntegrationType extends ApiGatewayIntegrationType,
-  Output,
-> = IntegrationType extends 'restApi'
-  ? APIGatewayProxyResult
-  : APIGatewayProxyResultV2<Output>;
-
-/**
- * The type of an ApiGateway handler. This is the actual version that will
- * be executed by the lambda, not the Swarmion inferred one.
- *
- * See https://docs.aws.amazon.com/lambda/latest/dg/typescript-handler.html.
- */
-export type ApiGatewayHandler<
-  IntegrationType extends ApiGatewayIntegrationType,
-  AuthorizerType extends ApiGatewayAuthorizerType,
-  Output,
-  AdditionalArgs extends unknown[] = never[],
-> = (
-  event: ApiGatewayEvent<IntegrationType, AuthorizerType>,
-  context: Context,
-  callback: HandlerCallback<IntegrationType>,
-  ...additionalArgs: AdditionalArgs
-) => Promise<ApiGatewayResult<IntegrationType, Output>>;
